вынес несколько переменных в const.js
