--- js/data.js
@@ -1,5 +1,0 @@
-//TODO: переименовать файл в CONST.js
-
-const COMMENTS_BATCH_LENGTH = 5;
-
-export { COMMENTS_BATCH_LENGTH};
