+++ js/load.js
@@ -1,3 +1,4 @@
+import { SERVER_ADDRESS } from './const.js';
 const createErrorMessage = document.querySelector('#error').content.querySelector('.error');
 const showError = () => {
   const newError = createErrorMessage.cloneNode(true);
@@ -9,7 +10,7 @@ const showError = () => {
 };
 
 
-export const getPosts = () => fetch('https://28.javascript.htmlacademy.pro/kekstagram/data') //TODO: вынести переменную
+export const getPosts = () => fetch(SERVER_ADDRESS) //TODO: вынести переменную
   .then((response) => {
     if (response.ok) {
       return response;
